design stuff, removing unused files
