+++ pages/_app.tsx
@@ -3,4 +3,5 @@ import type { AppProps } from 'next/app'
 
 export default function App({ Component, pageProps }: AppProps) {
   return <Component {...pageProps} />
+
 }
