+++ pages/components/faderView.tsx
@@ -1,6 +1,7 @@
 
 
 import { useEffect, useState } from "react"
+import {roboto} from '..'
 
 export default function FaderView(props:any){
    
@@ -47,17 +48,17 @@ export default function FaderView(props:any){
     
 
         return(
-        <div className="flex flex-col">
+        <div className={` flex flex-col`}>
     
-        <h1 className="p-2">{props.title}</h1>
+        <h1 className="p-2 text-black">{props.title}</h1>
     <hr className="relative ml-[85px] mr-[89px] top-[29px] shadow-md border-black"></hr>
 
-    <div className="flex justify-around p-2 bg-gray-200 rounded-md shadow-md">
+    <div className="flex justify-around p-2 bg-white rounded-md shadow-md">
     <div className="absolute">
         <div id="thumb" className={buttonAnimation}></div>
         </div>
 
-        <button className="p-2 px-4 mx-2 bg-white rounded-lg shadow-lg"onClick={()=>{
+        <button className="p-2 px-4 mx-2 rounded-lg shadow-lg bg-white text-black"onClick={()=>{
             if(faderVal === 0) return
             
             props.setFaderValue(props.faderValue-1)
@@ -66,31 +67,31 @@ export default function FaderView(props:any){
         <div className="flex items-center px-2">
 
       
-        <button className="px-3" onClick={()=>{
+        <button className=" text-black px-3" onClick={()=>{
             animateThumb(faderVal,0)
             setFaderVal(0)
         }}>|
         </button>
                    
-         <button className="px-3" onClick={()=>{
+         <button className=" text-black px-3" onClick={()=>{
             animateThumb(faderVal,1)
             setFaderVal(1)
         }}>|</button>
-         <button className="px-3" onClick={()=>{
+         <button className=" text-black px-3" onClick={()=>{
              animateThumb(faderVal,2)
              setFaderVal(2)
             }}>|</button>
          
-         <button className="px-3" onClick={()=>{
+         <button className=" text-black px-3" onClick={()=>{
              animateThumb(faderVal,3)
              setFaderVal(3)
             }}>|</button>
-         <button className="px-3" onClick={()=>{
+         <button className=" text-black px-3" onClick={()=>{
              animateThumb(faderVal,4)
              setFaderVal(4)
             }}>|</button>
             </div>
-        <button className="p-2 px-4 rounded-lg shadow-lg bg-white mx-2" onClick={()=>{
+        <button className="p-2 px-4 rounded-lg shadow-lg bg-white mx-2 text-black" onClick={()=>{
             if(faderVal === 4) return
             setFaderVal(faderVal+1)
             animateThumb(faderVal,faderVal+1)
