+++ pages/doll/DollPage.tsx
@@ -2,6 +2,7 @@ import FaderView from '../components/faderView'
 import { useRouter } from 'next/router'
 import { useEffect, useState } from 'react'
 import Image from 'next/image'
+import {roboto} from '..'
 
 
 
@@ -31,16 +32,16 @@ export default function DollPage(){
         setDollImgName(`/imgs_lela/0${busto+1}0${cintura+1}0${quadril+1}.jpeg`) 
     }
 
-    return (<div className="p-4 m-2 flex outline-1 rounded-lg shadow-md justify-between items-center max-w-lg">
+    return (<div className={`${roboto.className} flex outline-1 rounded-lg shadow-md justify-between items-center bg-white`}>
         <div className="flex shadow-none mr-2">
-            <div>
+            <div className='flex min-w-[10rem] max-w-[13rem]'>
         <Image width={600} height={1600} priority={true} alt={'a manequin that changes its sizes'} src={dollImgName}/>
             </div>
         </div>
         <div className='flex'>
-        <div className='flex flex-col p-2 bg-gray-100 rounded-lg shadow-sm'>
-        <h1 className="text-gray-800 pl-3 pt-5 pb-5 font-medium text-xl">
-            Descubra Seu Tamanho
+        <div className='flex flex-col p-2 bg-white rounded-lg shadow-sm'>
+        <h1 className="text-gray-800 pl-3 pt-5 pb-5 text-xl font-medium">
+            Ajuste conforme o formato do seu corpo
         </h1>
         {/* we can do some logic here to define what can i pass through */}
         <FaderView title='Busto' faderValue={busto} setFaderValue={(a:any)=>setBusto(a)} 
@@ -49,7 +50,7 @@ export default function DollPage(){
         ></FaderView>
         <FaderView title='Quadril' faderValue={quadril} setFaderValue={(c:any)=>setQuadril(c)}
         ></FaderView>
-        <button className="rounded-lg bg-gray-50 shadow-lg py-2 my-4 mt-10 " onClick={()=>router.push({pathname:'/result/refactor_resultPage', query:{dollImg:dollImgName, encodedImgUrl:encodedImgUrl,doll:true, imc:imc, busto:busto, cintura:cintura, quadril:quadril, categoria:categoria}})}>PRÓXIMO</button>
+        <button className="rounded-lg bg-black shadow-lg py-2 my-4 mt-10 text-white  " onClick={()=>router.push({pathname:'/result/resultPage', query:{dollImg:dollImgName, encodedImgUrl:encodedImgUrl,doll:true, imc:imc, busto:busto, cintura:cintura, quadril:quadril, categoria:categoria}})}>PRÓXIMO</button>
 
     </div>
