+++ pages/result/resultPage.tsx
@@ -1,559 +1,656 @@
-import Image from 'next/image'
-import { useRouter } from 'next/router'
-import { useEffect, useState } from 'react'
-import OtherSizesFactory from '../components/OtherSizesFactory'
-
+import { useRouter } from "next/router"
+import {useEffect,useState } from "react"
 import SugarBustoSVG from '../components/sugar/SugarBusto'
 import SugarCinturaSVG from '../components/sugar/SugarCintura'
 import SugarQuadrilSVG from '../components/sugar/SugarQuadril'
+import ChangeSizeButtons from "../components/ChangeSizeButtons"
+import { roboto } from ".."
+
+export default function ResultPage(){
 
 
-
-
-
-
-export default function ResultPage() {
     const router = useRouter()
-
-
-    const categoria = router.query.categoria
-
-
-    // const dollImg = router.query.dollImg
+    
     const encodedImgUrl = String(router.query.encodedImgUrl)
-    console.log('olha o encoded na result page,', encodedImgUrl)
-
-    const imc = Number(router.query.imc)
+    const categoria = String(router.query.categoria)
+    const edit = router.query.edit
+    const doll = router.query.doll
     
 
-    const busto = Number(router.query.busto)
-    const cintura = Number(router.query.cintura)
-    const quadril = Number(router.query.quadril)
-
-    const bustoCm = Number(router.query.bustoCm)
-    const cinturaCm = Number(router.query.cinturaCm)
-    const quadrilCm = Number(router.query.quadrilCm)
-
-
-
+    
     const [sizeTop, setSizeTop] = useState<string>('')
     const [sizeBottom, setSizeBottom] = useState<string>('')
     const [sizeWhole, setSizeWhole] = useState<string>('')
 
-    const [finalSize,setFinalSize] = useState<string>('')
 
-    //sugar states
+
+
+    
     const [bustoDescription, setBustoDescription] = useState<string>('')
     const [bustoColor, setBustoColor] = useState<string>('')
+    const [editBusto, setEditBusto] = useState<string | number>('')
     const [cinturaDescription, setCinturaDescription] = useState<string>('')
+    const [editCintura, setEditCintura] = useState<string | number>('')
     const [cinturaColor, setCinturaColor] = useState<string>('')
     const [quadrilDescription, setQuadrilDescription] = useState<string>('')
+    const [editQuadril, setEditQuadril] = useState<string | number>('')
     const [quadrilColor, setQuadrilColor] = useState<string>('')
+    
+  
 
-    const [editBusto, setEditBusto] = useState<string | number>()
-    const [editCintura, setEditCintura] = useState<string | number>()
-    const [editQuadril, setEditQuadril] = useState<string | number>()
-
-    //boolean values to setup the useEffect
-    const doll = router.query.doll
-    const edit = router.query.edit
-    useEffect(() => {
-        //dollpage
-        if (doll){
-            if(imc){
-                console.warn('can it enter?')
-            if(busto && cintura && quadril || busto===0 || cintura===0 || quadril===0) {
-            chooseSizeDoll(imc, busto, cintura, quadril)
-            }
-        }
-    }
-
-        //exactMeasures
-        if (edit){
-            if(router.query.sizeTop){        
-                if(router.query.sizeBottom && router.query.sizeWhole) {
-                    setter(String(router.query.sizeTop), String(router.query.sizeBottom), String(router.query.sizeWhole), String(router.query.bustoDescription), String(router.query.bustoColor), String(router.query.cinturaDescription), String(router.query.cinturaColor), String(router.query.quadrilDescription), String(router.query.quadrilColor))
-                } 
-            }
-        }
-
-    }, [router.query])
-
-    // change this name
-    const setter = (sizeTop:string, sizeBottom:string, sizeWhole:string, bustoDesc:string, bustoColorEdit:string, cinturaDesc:string, cinturaColorEdit:string, quadrilDesc:string, quadrilColorEdit:string) => {
-        console.log('inside setter', { sizeTop, sizeBottom, sizeWhole })
-        setSizeTop(sizeTop);
-        setSizeBottom(sizeBottom);
-        setSizeWhole(sizeWhole);
-        setEditBusto(bustoCm)
-        setEditCintura(cinturaCm)
-        setEditQuadril(quadrilCm)
-        setBustoDescription(bustoDesc)
-        setBustoColor(bustoColorEdit)
-        setCinturaDescription(cinturaDesc)
-        setCinturaColor(cinturaColorEdit)
-        setQuadrilDescription(quadrilDesc)
-        setQuadrilColor(quadrilColorEdit)
-    }
+    const [factoryResultState,setFactoryResultState] = useState<any>()    
 
 
 
+    const imc = Number(router.query.imc)
+    const bustoDoll = Number(router.query.busto)
+    const cinturaDoll = Number(router.query.cintura)
+    const quadrilDoll = Number(router.query.quadril)
+
+
+
+
+
+    const allPossibleCategorias = ['Camisa', 'Calça', 'Vestidos']
+    const resultNameCategorias = ['sizeTop', 'sizeBottom', 'sizeWhole']
+    const sizes = [sizeTop,sizeBottom,sizeWhole]
+    const functionsDollObj = [(imc: number,busto: number,cintura: number,quadril: number)=> chooseBestCamisaDoll(imc,busto,cintura,quadril), (imc: number,busto: number,cintura: number,quadril: number)=> chooseBestCalçaDoll(imc,busto,cintura,quadril), (imc: number,busto: number,cintura: number,quadril: number)=>chooseBestVestidoDoll(imc,busto,cintura,quadril) ]
+    
+    const imcRanges = [14,20,23.5,28.7,30,32.5,35.5]
     const allSizesNames = ['PP', 'P', 'M', 'G', 'GG', 'XG']
     const allDescriptions = ['Largo', 'Folgado', 'Levemente folgado', 'Ideal', 'Levemente Justo', 'Justo', 'Apertado']
     const allColors = ['stroke-red-500', 'stroke-yellow-500', 'stroke-green-300', 'stroke-green-500','stroke-green-300', 'stroke-yellow-500','stroke-red-500']
-    const imcRanges = [14,20,23.5,28.7,30,32.5,35.5]
+
     const PP = { busto: { min: 75, med: 80.5, max: 86 }, cintura: { min: 65, med: 67.5, max: 70 }, quadril: { min: 92, med: 95, max: 98 } }
     const P = { busto: { min: 87, med: 91.5, max: 96 }, cintura: { min: 70, med: 73, max: 76 }, quadril: { min: 99, med: 102, max: 105 } }
     const M = { busto: { min: 97, med: 99.5, max: 102 }, cintura: { min: 77, med: 79, max: 81 }, quadril: { min: 106, med: 108.5, max: 111 } }
     const G = { busto: { min: 103, med: 106, max: 109 }, cintura: { min: 82, med: 84.5, max: 87 }, quadril: { min: 112, med: 115, max: 118 } }
     const GG = { busto: { min: 110, med: 114.5, max: 119 }, cintura: { min: 88, med: 90.5, max: 93 }, quadril: { min: 119, med: 121.5, max: 124 } }
     const XG = { busto: { min: 120, med: 125, max: 130 }, cintura: { min: 94, med: 96.5, max: 99 }, quadril: { min: 125, med: 127.5, max: 130 } }
-
-    const allSizes=[PP,P,M,G,GG,XG]
+    const allSizes = [PP,P,M,G,GG,XG]
     
-    const setSizeTopPlusOne = (sizePlusOne: string, apxBusto: number,  apxCintura: number, apxQuadril: number)=>{
-        setBustoDescription(allDescriptions[busto-2])
-        setBustoColor(allColors[busto-2])
-        setEditBusto(apxBusto)
-        setCinturaDescription(allDescriptions[cintura])
-        setCinturaColor(allColors[cintura])
-        setEditCintura(apxCintura)
-        setQuadrilDescription(allDescriptions[quadril])
-        setQuadrilColor(allColors[quadril])
-        setEditQuadril(apxQuadril)
-        
-        setSizeTop(sizePlusOne)
+    const factoryInfo = {
+        sizeTop,
+        sizeBottom,
+        sizeWhole,
+        editBusto,
+        editCintura,
+        editQuadril,
+        bustoDescription:allDescriptions.indexOf(bustoDescription),
+        bustoColor:allDescriptions.indexOf(bustoDescription),
+        cinturaDescription:allDescriptions.indexOf(cinturaDescription),
+        cinturaColor:allDescriptions.indexOf(cinturaDescription),
+        quadrilDescription:allDescriptions.indexOf(quadrilDescription),
+        quadrilColor:allDescriptions.indexOf(quadrilDescription),
     }
-    const setSizeTopMinusOne = (sizeMinusOne: string, apxBusto: number,apxCintura: number,apxQuadril: number) =>{
-        setBustoDescription(allDescriptions[busto+5])
-        setBustoColor(allColors[busto+5])
-        setEditBusto(apxBusto)
-        setCinturaDescription(allDescriptions[cintura])
-        setCinturaColor(allColors[cintura])
-        setEditCintura(apxCintura)
-        setQuadrilDescription(allDescriptions[quadril])
-        setQuadrilColor(allColors[quadril])
-        setEditQuadril(apxQuadril)
-        
-        setSizeTop(sizeMinusOne)
-        
-    }
-    const setSizeBottomAndWholePlusOne = (sizePlusOne: string, apxBusto: number,  apxCintura: number, apxQuadril: number)=>{
-        setBustoDescription(allDescriptions[busto])
-        setBustoColor(allColors[busto])
-        setEditBusto(apxBusto)
-        setCinturaDescription(allDescriptions[cintura-2])
-        setCinturaColor(allColors[cintura-2])
-        setEditCintura(apxCintura)
-        setQuadrilDescription(allDescriptions[quadril-2])
-        setQuadrilColor(allColors[quadril-2])
-        setEditQuadril(apxQuadril)
-        
-        setSizeBottom(sizePlusOne)
-        setSizeWhole(sizePlusOne)
-    }
-    const setSizeBottomAndWholeMinusOne = (sizeMinusOne: string, apxBusto: number, apxCintura: number,apxQuadril: number) =>{
-        //if whole || bottom?
-        setBustoDescription(allDescriptions[busto])
-        setBustoColor(allColors[busto])
-        setEditBusto(apxBusto)
-        setCinturaDescription(allDescriptions[cintura+2])
-        setCinturaColor(allColors[cintura+2])
-        setEditCintura(apxCintura)
-        setQuadrilDescription(allDescriptions[quadril+2])
-        setQuadrilColor(allColors[quadril+2])
-        setEditQuadril(apxQuadril)
-        
-        setSizeBottom(sizeMinusOne)
-        setSizeWhole(sizeMinusOne)
-    }
-
-    const imcCheck = (imc: number) =>{
+    const imcCheck = (imc:number) =>{
         if (imc < 14 || imc > 35.5) {
             router.push({ pathname: '/medidas/NotFound', query: { encodedImgUrl: encodedImgUrl } })
             //go to the Nao encontramos seu tamanho por favor -> Editar Medidas
         }
         
     }
-
-    //change names
-    const chooseSizeDoll = (imc: number, busto: number, cintura: number, quadril: number) => {
-        console.log('chooseSizeDoll')
+    //doll
+    const chooseBestCamisaDoll = (imc:number,busto:number,cintura:number,quadril:number) =>{
+        
         imcCheck(imc)
-        imcRanges.forEach((v,i,arr)=>{
-            //PP
-         if(i===0){
-            console.log('your size is PP')
-             const bustosPP = [PP.busto.min, PP.busto.min+1, PP.busto.med+1, PP.busto.max, P.busto.min]
-             const quadrisPP = [PP.quadril.min, PP.quadril.min+1, PP.quadril.med+1, PP.quadril.max, P.quadril.min]
-             const cinturasPP = [PP.cintura.min, PP.cintura.min+1, PP.cintura.med+1, PP.cintura.max, P.cintura.min]
+        console.log('inside chooseBestCamisa')
+        const setSizePlusOne = (sizePlusOne:string,apxBusto:number,apxCintura:number,apxQuadril:number) =>{
+            setBustoDescription(allDescriptions[busto-2])
+            setBustoColor(allColors[busto-2])
+            setEditBusto(apxBusto)
+            setCinturaDescription(allDescriptions[cintura])
+            setCinturaColor(allColors[cintura])
+            setEditCintura(apxCintura)
+            setQuadrilDescription(allDescriptions[quadril])
+            setQuadrilColor(allColors[quadril])
+            setEditQuadril(apxQuadril)
+            setSizeTop(sizePlusOne)
+            return
+        }
+        const setSizeMinusOne = (sizeMinusOne:string,apxBusto:number,apxCintura:number,apxQuadril:number) =>{
+            setBustoDescription(allDescriptions[busto+5])
+            setBustoColor(allColors[busto+5])
+            setEditBusto(apxBusto)
+            setCinturaDescription(allDescriptions[cintura+2])
+            setCinturaColor(allColors[cintura+2])
+            setEditCintura(apxCintura)
+            setQuadrilDescription(allDescriptions[quadril+2])
+            setQuadrilColor(allColors[quadril+2])
+            setEditQuadril(apxQuadril)
+            setSizeTop(sizeMinusOne)
+            return
+        }
+
+        imcRanges.forEach((v,i,imcRanges)=>{
+            if(i===imcRanges.length && i ===imcRanges.length-1)return
+             //PP
+            if(i===0 && imc>imcRanges[i] && imc<imcRanges[i+1]){
+            // console.log('your size is PP')
+             const bustos = [allSizes[i].busto.min, allSizes[i].busto.min+1, allSizes[i].busto.med+1, allSizes[i].busto.max, allSizes[i+1].busto.min]
+             const quadris = [allSizes[i].quadril.min, allSizes[i].quadril.min+1, allSizes[i].quadril.med+1, allSizes[i].quadril.max, allSizes[i+1].quadril.min]
+             const cinturas = [allSizes[i].cintura.min, allSizes[i].cintura.min+1, allSizes[i].cintura.med+1, allSizes[i].cintura.max, allSizes[i+1].cintura.min]
              
-             const blockValue = 0
-             const defaultSize = allSizesNames[blockValue]
-             const sizePlusOne = allSizesNames[blockValue + 1]
-             console.warn('is 0?')
-             console.warn(busto, typeof busto,' busto log warn')
-             
-             if (busto==0) {
-                 //folgado 
-                 setBustoDescription(allDescriptions[busto+1])
-                 setBustoColor(allColors[busto+1])
-                 setSizeTop(defaultSize)
-                 setEditBusto(bustosPP[blockValue+busto])
-             }
-             if (busto == 1) {
-                 //levemente folgado
-                 setBustoDescription(allDescriptions[busto+1])
-                 setBustoColor(allColors[busto+1])
-                 setSizeTop(defaultSize)
-                 setEditBusto(bustosPP[blockValue+busto])
-             }
-             if (busto == 2) {
-                 //levemente justo
-                 setBustoDescription(allDescriptions[busto+1])
-                 setBustoColor(allColors[busto+1])
-                 setSizeTop(defaultSize)
-                 setEditBusto(bustosPP[blockValue+busto])
-             }
-             if (busto == 3) {
-                 //justo
-                 setBustoDescription(allDescriptions[busto+1])
-                 setBustoColor(allColors[busto+1])
-                 setSizeTop(defaultSize)
-                 setEditBusto(bustosPP[blockValue+busto])
-             }
-             if (busto == 4) {
-                 //folgado
-                return setSizeTopPlusOne(sizePlusOne,bustosPP[blockValue+busto],cinturasPP[blockValue+cintura],quadrisPP[blockValue+quadril])
-             }
-             if (cintura == 0) {
-                 setCinturaDescription(allDescriptions[cintura+1])
-                 setCinturaColor(allColors[cintura+1])
-                 setEditCintura(cinturasPP[cintura])
-             }
-             if (cintura == 1) {
-                 setCinturaDescription(allDescriptions[cintura+1])
-                 setCinturaColor(allColors[cintura+1])
-                 setEditCintura(cinturasPP[cintura])
-             }
-             if (cintura == 2) {
-                 setCinturaDescription(allDescriptions[cintura+1])
-                 setCinturaColor(allColors[cintura+1])
-                 setEditCintura(cinturasPP[cintura])
-             }
-             if (cintura == 3) {
-                 setCinturaDescription(allDescriptions[cintura+1])
-                 setCinturaColor(allColors[cintura+1])
-                 setEditCintura(cinturasPP[cintura])
-             }
-             if (cintura == 4) {
-                 setCinturaDescription(allDescriptions[cintura+1])
-                 setCinturaColor(allColors[cintura+1])
-                 setEditCintura(cinturasPP[cintura])
-             }
-             if (quadril == 0) {
-                 setQuadrilDescription(allDescriptions[quadril+1])
-                 setQuadrilColor(allColors[quadril+1])
-                 setSizeBottom(defaultSize)
-                 setSizeWhole(defaultSize)
-                 setEditQuadril(quadrisPP[quadril])
-             }
- 
-             if (quadril == 1) {
-                 setQuadrilDescription(allDescriptions[quadril+1])
-                 setQuadrilColor(allColors[quadril+1])
-                 setSizeBottom(defaultSize)
-                 setSizeWhole(defaultSize)
-                 setEditQuadril(quadrisPP[quadril])
-             }
-             if (quadril == 2) {
-                 setQuadrilDescription(allDescriptions[quadril+1])
-                 setQuadrilColor(allColors[quadril+1])
-                 setSizeBottom(defaultSize)
-                 setSizeWhole(defaultSize)
-                 setEditQuadril(quadrisPP[quadril])
-             }
- 
-             if (quadril == 3) {
-                 setQuadrilDescription(allDescriptions[quadril+1])
-                 setQuadrilColor(allColors[quadril+1])
-                 setSizeBottom(defaultSize)
-                 setSizeWhole(defaultSize)
-                 setEditQuadril(quadrisPP[quadril])
-             }
-             if (quadril == 4) {
-                return setSizeBottomAndWholePlusOne(sizePlusOne,bustosPP[blockValue+busto],cinturasPP[blockValue+cintura],quadrisPP[blockValue+quadril])
-             }
-         }
-         //XG ou ultimo tamanho
-         if(i==arr.length){
              const blockValue = i
              const defaultSize = allSizesNames[blockValue]
-             const sizeMinusOne = allSizesNames[blockValue - 1]
-             const bustos = [allSizes[blockValue-1].busto.max,allSizes[blockValue].busto.min,allSizes[blockValue].busto.med-1,allSizes[blockValue].busto.med+1, allSizes[blockValue].busto.max]
-             const cinturas = [allSizes[blockValue-1].cintura.max,allSizes[blockValue].cintura.min,allSizes[blockValue].cintura.med-1,allSizes[blockValue].cintura.med+1, allSizes[blockValue].cintura.max]
-             const quadris = [allSizes[blockValue-1].quadril.max,allSizes[blockValue].quadril.min,allSizes[blockValue].quadril.med-1,allSizes[blockValue].quadril.med+1, allSizes[blockValue].quadril.max]
-             if (busto == 0) {
-                 //folgado
- 
-                 return setSizeTopMinusOne(sizeMinusOne, bustos[busto], cinturas[cintura], quadris[quadril])
-                 // setBustoDescription(allDescriptions[busto+5])
-                 // setBustoColor(allColors[busto+5])
- 
-                 
-             }
-             if (busto == 1) {
-                 //levemente folgado
+             const sizePlusOne = allSizesNames[blockValue + 1]
+            //  console.warn('is 0?')
+            //  console.warn(busto, typeof busto,' busto log warn')
+        
+             if (busto>0 && busto<4) {
                  setBustoDescription(allDescriptions[busto+1])
                  setBustoColor(allColors[busto+1])
                  setSizeTop(defaultSize)
-                 setEditBusto(bustos[busto])
+                 setEditBusto(bustos[blockValue+busto])
              }
-             if (busto == 2) {
-                 //Ideal
-                 setBustoDescription(allDescriptions[busto+1])
-                 setBustoColor(allColors[busto+1])
-                 setSizeTop(defaultSize)
-                 setEditBusto(bustos[busto])
-             }
-             if (busto == 3) {
-                 //Levemente Justo
-                 setBustoDescription(allDescriptions[busto+1])
-                 setBustoColor(allColors[busto+1])
-                 setSizeTop(defaultSize)
-                 setEditBusto(bustos[busto])
-             }
+        
              if (busto == 4) {
-                 setBustoDescription(allDescriptions[busto+1])
-                 setBustoColor(allColors[busto+1])
-                 setSizeTop(defaultSize)
-                 setEditBusto(bustos[busto])
-                 
-                 
-                 //levemente folgado
-                 // setBustoColor(allDescriptions[busto-2])
-                 // setBustoDescription(allDescriptions[busto-2])
+                return setSizePlusOne(sizePlusOne,bustos[blockValue+busto],cinturas[blockValue+cintura],quadris[blockValue+quadril])
              }
-             if (cintura == 0) {
-                 setCinturaDescription(allDescriptions[cintura+1])
-                 setCinturaColor(allColors[cintura+1])
-                 setEditCintura(cinturas[cintura])
-             }
-             if (cintura == 1) {
-                 setCinturaDescription(allDescriptions[cintura+1])
-                 setCinturaColor(allColors[cintura+1])
-                 setEditCintura(cinturas[cintura])
-             }
-             if (cintura == 2) {
-                 setCinturaDescription(allDescriptions[cintura+1])
-                 setCinturaColor(allColors[cintura+1])
-                 setEditCintura(cinturas[cintura])
-             }
-             if (cintura == 3) {
-                 setCinturaDescription(allDescriptions[cintura+1])
-                 setCinturaColor(allColors[cintura+1])
-                 setEditCintura(cinturas[cintura])
-             }
-             if (cintura == 4) {
-                 setCinturaDescription(allDescriptions[cintura+1])
-                 setCinturaColor(allColors[cintura+1])
-                 setEditCintura(cinturas[cintura])
-             }
-             if (quadril == 0) {
-                 // setQuadrilDescription('Folgado')
-                 // setQuadrilColor('stroke-red-500')
-                 // setSizeBottom(sizeMinusOne)
-                 // setSizeWhole(sizeMinusOne)
-                 // setEditQuadril(PP.quadril.max)
-                return setSizeBottomAndWholeMinusOne(sizeMinusOne, bustos[busto], cinturas[cintura], quadris[quadril])
-                 
-             }
- 
-             if (quadril == 1) {
- 
-                 setQuadrilDescription(allDescriptions[quadril+1])
-                 setQuadrilColor(allColors[quadril+1])
-                 setSizeBottom(defaultSize)
-                 setSizeWhole(defaultSize)
-                 setEditQuadril(quadris[quadril])
-             }
-             if (quadril == 2) {
-                 setQuadrilDescription(allDescriptions[quadril+1])
-                 setQuadrilColor(allColors[quadril+1])
-                 setSizeBottom(defaultSize)
-                 setSizeWhole(defaultSize)
-                 setEditQuadril(quadris[quadril])
-             }
- 
-             if (quadril == 3) {
-                 setQuadrilDescription(allDescriptions[quadril+1])
-                 setQuadrilColor(allColors[quadril+1])
-                 setSizeBottom(defaultSize)
-                 setSizeWhole(defaultSize)
-                 setEditQuadril(quadris[quadril])
-             }
-             if (quadril == 4) {
-                 setQuadrilDescription(allDescriptions[quadril+1])
-                 setQuadrilColor(allColors[quadril+1])
-                 setSizeBottom(defaultSize)
-                 setSizeWhole(defaultSize)
-                 setEditQuadril(quadris[quadril])
-             }
+             
+                setCinturaDescription(allDescriptions[cintura+1])
+                setCinturaColor(allColors[cintura+1])
+                setEditCintura(cinturas[cintura])
+                setQuadrilDescription(allDescriptions[quadril+1])
+                setQuadrilColor(allColors[quadril+1])
+                setEditQuadril(quadris[quadril])
+                return        
          }
-         //P ao GG
-         if(i!==arr.length && i!==0){
-             if(imc>arr[i] && imc<arr[i+1]){
-                 const blockValue = i
-                 const defaultSize = allSizesNames[blockValue]
-                 const sizePlusOne = allSizesNames[blockValue + 1]
-                 const sizeMinusOne = allSizesNames[blockValue - 1]
-                 const bustos = [allSizes[blockValue-1].busto.max,allSizes[blockValue].busto.min,allSizes[blockValue].busto.med,allSizes[blockValue].busto.max, allSizes[blockValue+1].busto.min]
-                 const cinturas = [allSizes[blockValue-1].cintura.max,allSizes[blockValue].cintura.min,allSizes[blockValue].cintura.med,allSizes[blockValue].cintura.max, allSizes[blockValue+1].cintura.min]
-                 const quadris = [allSizes[blockValue-1].quadril.max,allSizes[blockValue].quadril.min,allSizes[blockValue].quadril.med,allSizes[blockValue].quadril.max, allSizes[blockValue+1].quadril.min]
-                 if (busto == 0) {
-                     //folgado
-     
-                     return setSizeTopMinusOne(sizeMinusOne, bustos[busto], cinturas[cintura], quadris[quadril])
-                     // setBustoDescription(allDescriptions[busto+5])
-                     // setBustoColor(allColors[busto+5])
-     
-                     
-                 }
-                 if (busto == 1) {
-                     //levemente folgado
-                     setBustoDescription(allDescriptions[busto+1])
-                     setBustoColor(allColors[busto+1])
-                     setSizeTop(defaultSize)
-                     setEditBusto(bustos[busto])
-                 }
-                 if (busto == 2) {
-                     //Ideal
-                     setBustoDescription(allDescriptions[busto+1])
-                     setBustoColor(allColors[busto+1])
-                     setSizeTop(defaultSize)
-                     setEditBusto(bustos[busto])
-                 }
-                 if (busto == 3) {
-                     //Levemente Justo
-                     setBustoDescription(allDescriptions[busto+1])
-                     setBustoColor(allColors[busto+1])
-                     setSizeTop(defaultSize)
-                     setEditBusto(bustos[busto])
-                 }
-                 if (busto == 4) {
-                     
-                     setSizeTopPlusOne(sizePlusOne, bustos[busto], cinturas[cintura], quadris[quadril])
-                     //levemente folgado
-                     // setBustoColor(allDescriptions[busto-2])
-                     // setBustoDescription(allDescriptions[busto-2])
-                 }
-                 if (cintura == 0) {
-                     setCinturaDescription(allDescriptions[cintura+1])
-                     setCinturaColor(allColors[cintura+1])
-                     setEditCintura(cinturas[cintura])
-                 }
-                 if (cintura == 1) {
-                     setCinturaDescription(allDescriptions[cintura+1])
-                     setCinturaColor(allColors[cintura+1])
-                     setEditCintura(cinturas[cintura])
-                 }
-                 if (cintura == 2) {
-                     setCinturaDescription(allDescriptions[cintura+1])
-                     setCinturaColor(allColors[cintura+1])
-                     setEditCintura(cinturas[cintura])
-                 }
-                 if (cintura == 3) {
-                     setCinturaDescription(allDescriptions[cintura+1])
-                     setCinturaColor(allColors[cintura+1])
-                     setEditCintura(cinturas[cintura])
-                 }
-                 if (cintura == 4) {
-                     setCinturaDescription(allDescriptions[cintura+1])
-                     setCinturaColor(allColors[cintura+1])
-                     setEditCintura(cinturas[cintura])
-                 }
-                 if (quadril == 0) {
-                     // setQuadrilDescription('Folgado')
-                     // setQuadrilColor('stroke-red-500')
-                     // setSizeBottom(sizeMinusOne)
-                     // setSizeWhole(sizeMinusOne)
-                     // setEditQuadril(PP.quadril.max)
-                     setSizeBottomAndWholeMinusOne(sizePlusOne, bustos[busto], cinturas[cintura], quadris[quadril])
-                     
-                 }
-     
-                 if (quadril == 1) {
-     
-                     setQuadrilDescription(allDescriptions[quadril+1])
-                     setQuadrilColor(allColors[quadril+1])
-                     setSizeBottom(defaultSize)
-                     setSizeWhole(defaultSize)
-                     setEditQuadril(quadris[quadril])
-                 }
-                 if (quadril == 2) {
-                     setQuadrilDescription(allDescriptions[quadril+1])
-                     setQuadrilColor(allColors[quadril+1])
-                     setSizeBottom(defaultSize)
-                     setSizeWhole(defaultSize)
-                     setEditQuadril(quadris[quadril])
-                 }
-     
-                 if (quadril == 3) {
-                     setQuadrilDescription(allDescriptions[quadril+1])
-                     setQuadrilColor(allColors[quadril+1])
-                     setSizeBottom(defaultSize)
-                     setSizeWhole(defaultSize)
-                     setEditQuadril(quadris[quadril])
-                 }
-                 if (quadril == 4) {
-     
-                     setSizeBottomAndWholePlusOne(sizePlusOne, bustos[busto], cinturas[cintura], quadris[quadril])
-                 }
-                     }
-             }
-         }) 
+         //XG
+         if(i===imcRanges.length-2 && imc>imcRanges[i] && imc<imcRanges[i+1]){
+            const blockValue = i
+            const defaultSize = allSizesNames[blockValue]
+            const sizeMinusOne = allSizesNames[blockValue - 1]
+            const bustos = [allSizes[blockValue-1].busto.max,allSizes[blockValue].busto.min,allSizes[blockValue].busto.med-1,allSizes[blockValue].busto.med+1, allSizes[blockValue].busto.max]
+            const cinturas = [allSizes[blockValue-1].cintura.max,allSizes[blockValue].cintura.min,allSizes[blockValue].cintura.med-1,allSizes[blockValue].cintura.med+1, allSizes[blockValue].cintura.max]
+            const quadris = [allSizes[blockValue-1].quadril.max,allSizes[blockValue].quadril.min,allSizes[blockValue].quadril.med-1,allSizes[blockValue].quadril.med+1, allSizes[blockValue].quadril.max]
+            if (busto == 0) {
+                return setSizeMinusOne(sizeMinusOne, bustos[busto], cinturas[cintura], quadris[quadril])
+            }
+            if (busto >= 1 && busto<=4) {
+                setBustoDescription(allDescriptions[busto+1])
+                setBustoColor(allColors[busto+1])
+                setSizeTop(defaultSize)
+                setEditBusto(bustos[busto])
+            }
+            
+                setCinturaDescription(allDescriptions[cintura+1])
+                setCinturaColor(allColors[cintura+1])
+                setEditCintura(cinturas[cintura])
+                setQuadrilDescription(allDescriptions[quadril+1])
+                setQuadrilColor(allColors[quadril+1])
+                setEditQuadril(quadris[quadril])
+            
+        }
+        //P ao GG
+        if(i!==imcRanges.length-1 && i!==imcRanges.length-2  && i!==0){
+            //P ao GG
+            if(imc>imcRanges[i] && imc<imcRanges[i+1]){
+                
+                const blockValue = i
+                const defaultSize = allSizesNames[blockValue]
+                const sizePlusOne = allSizesNames[blockValue + 1]
+                const sizeMinusOne = allSizesNames[blockValue - 1]
+                const bustos = [allSizes[blockValue-1].busto.max,allSizes[blockValue].busto.min,allSizes[blockValue].busto.med,allSizes[blockValue].busto.max, allSizes[blockValue+1].busto.min]
+                const cinturas = [allSizes[blockValue-1].cintura.max,allSizes[blockValue].cintura.min,allSizes[blockValue].cintura.med,allSizes[blockValue].cintura.max, allSizes[blockValue+1].cintura.min]
+                const quadris = [allSizes[blockValue-1].quadril.max,allSizes[blockValue].quadril.min,allSizes[blockValue].quadril.med,allSizes[blockValue].quadril.max, allSizes[blockValue+1].quadril.min]
+                if (busto == 0) {
+                     return setSizeMinusOne(sizeMinusOne, bustos[busto], cinturas[cintura], quadris[quadril])                 
+                }
+                if (busto >= 1 && busto<=3 ) {
+                    //levemente folgado
+                    setBustoDescription(allDescriptions[busto+1])
+                    setBustoColor(allColors[busto+1])
+                    setSizeTop(defaultSize)
+                    setEditBusto(bustos[busto])
+                }           
+                if (busto == 4) {
+                    
+                   return setSizePlusOne(sizePlusOne, bustos[busto], cinturas[cintura], quadris[quadril])
+                }
+              
+                  setCinturaDescription(allDescriptions[cintura+1])
+                  setCinturaColor(allColors[cintura+1])
+                  setEditCintura(cinturas[cintura])
+                  setQuadrilDescription(allDescriptions[quadril+1])
+                  setQuadrilColor(allColors[quadril+1])
+                  setEditQuadril(quadris[quadril])
+                  return
+                }
+            }
+        })
+
+
     }
-    console.log(router.query, 'router query', busto,'=busto tipo =' ,typeof busto,'inrouter is', router.query.busto,)
-    if(!router.isFallback)return (<div className='p-4 m-2 flex justify-start outline-1 rounded-lg shadow-md items-center max-w-lg'>
-        <div className=' px-2'>
-            <div className='max-w-xs'> 
-            coisa a imagen
-               {/* change structure and names here */}
-             <SugarBustoSVG bustoColor={bustoColor} />
-             <SugarCinturaSVG cinturaColor={cinturaColor} />
-             <SugarQuadrilSVG quadrilColor={quadrilColor} />
-             {/* change structure and names here */}
-            <img alt={'doll std img'} src='/imgs_lela/030303.jpeg'/>
-            {/* map allSizes to buttons, and uses a factory that fabricates an object that is the full result, and will interact here */}
-            {/* {allSizes.map((v,i,arr)=>{
-                return <button className="rounded-lg bg-gray-50 shadow-lg py-2 mx-2">{v}</button>
-            })} */}
-            <OtherSizesFactory></OtherSizesFactory>
-            </div>
-        </div>
-        <div className=' w-96 ml-1 pb-6'>
-            {/* change struct here */}
-        {/* <h1 className="text-gray-800 font-medium text-xl">
-                 Sua melhor opção de tamanho para tops ou blusas : {sizeTop}, shorts e calças {sizeBottom}, vestidos etc {sizeWhole}
-            </h1> */}
-            {categoria==='Camisa'? <h1 className='text-gray-800 font font-medium text-xl'> Sua opção de {categoria} é {sizeTop}</h1>: null}
-            {categoria==="Calça"? <h1 className='text-gray-800 font font-medium text-xl'> Sua opção de {categoria} é {sizeBottom}</h1>: null}
-            {categoria==="Vestido"? <h1 className='text-gray-800 font font-medium text-xl'> Sua opção de {categoria} é {sizeWhole}</h1>: null}
-            {/* change struct here */}
-            <div className='mt-10'>{bustoDescription}</div>
-            <div className='mt-8'>{cinturaDescription}</div>
-            <div className='mt-14'>{quadrilDescription}</div>
+    const chooseBestCalçaDoll = (imc:number,busto:number,cintura:number,quadril:number) =>{   
+        imcCheck(imc)
+        const setSizePlusOne = (sizePlusOne:string,apxBusto:number,apxCintura:number,apxQuadril:number) =>{
+            setBustoDescription(allDescriptions[busto-2])
+            setBustoColor(allColors[busto-2])
+            setEditBusto(apxBusto)
+            setCinturaDescription(allDescriptions[cintura])
+            setCinturaColor(allColors[cintura])
+            setEditCintura(apxCintura)
+            setQuadrilDescription(allDescriptions[quadril])
+            setQuadrilColor(allColors[quadril])
+            setEditQuadril(apxQuadril)
+            setSizeBottom(sizePlusOne)
+        }
+        const setSizeMinusOne = (sizeMinusOne:string,apxBusto:number,apxCintura:number,apxQuadril:number) =>{
+            setBustoDescription(allDescriptions[busto+5])
+            setBustoColor(allColors[busto+5])
+            setEditBusto(apxBusto)
+            setCinturaDescription(allDescriptions[cintura])
+            setCinturaColor(allColors[cintura])
+            setEditCintura(apxCintura)
+            setQuadrilDescription(allDescriptions[quadril])
+            setQuadrilColor(allColors[quadril])
+            setEditQuadril(apxQuadril)
+            setSizeBottom(sizeMinusOne)
+        }
 
-        <div className='flex flex-col justify-start mt-8'>
-             <button className="rounded-lg bg-gray-50 shadow-lg py-2 my-2 " onClick={() => router.push({ pathname: `${window.location.origin}/${encodeURIComponent(encodedImgUrl)}`, query:{categoria:categoria}})}>Reiniciar</button>
-            {doll ? <button className="rounded-lg bg-gray-50 shadow-lg py-2 my-2 " onClick={() => router.push({ pathname: `${window.location.origin}/doll/DollPage`, query: { imc: imc, encodedImgUrl: encodedImgUrl, categoria:categoria } })}>Voltar</button> : null}
-            <button className="rounded-lg bg-gray-50 shadow-lg py-2 my-2 " onClick={() => router.push({ pathname: `${window.location.origin}/medidas/EditarMedidas`, query: { imc: imc, editBusto: editBusto, editCintura: editCintura, editQuadril: editQuadril, encodedImgUrl: encodedImgUrl, categoria:categoria } })}>Editar medidas</button>
-             {/* <button className="rounded-lg bg-gray-50 shadow-lg py-2 my-2 " onClick={()=>router.push({pathname:'/medidas/NotFound', query:{imc:imc, encodedImgUrl:encodedImgUrl}})}>NotFound</button> */}
-             <a className='rounded-lg bg-gray-50 shadow-lg py-2 my-2 text-center' href='/files/fita-metrica-eufloria.pdf' target={"_blank"} rel='noopener noreferrer'>
-               Fita Metrica
-                </a>
-            </div>
+        imcRanges.forEach((v,i,arr)=>{
+            if(i===imcRanges.length && i ===imcRanges.length-1)return
+            //PP
+        if(i===0 && imc>imcRanges[i] && imc<imcRanges[i+1]){
+           console.log('your size is PP')
+            const bustos = [allSizes[i].busto.min, allSizes[i].busto.min+1, allSizes[i].busto.med+1, allSizes[i].busto.max, allSizes[i+1].busto.min]
+            const quadris = [allSizes[i].quadril.min, allSizes[i].quadril.min+1, allSizes[i].quadril.med+1, allSizes[i].quadril.max, allSizes[i+1].quadril.min]
+            const cinturas = [allSizes[i].cintura.min, allSizes[i].cintura.min+1, allSizes[i].cintura.med+1, allSizes[i].cintura.max, allSizes[i+1].cintura.min]
+            
+            const blockValue = i
+            const defaultSize = allSizesNames[blockValue]
+            const sizePlusOne = allSizesNames[blockValue + 1]
+            console.warn('is 0?')
+            console.warn(busto, typeof busto,' busto log warn')
+            
+            if (quadril>=0 && quadril<4) {
+                setQuadrilDescription(allDescriptions[quadril+1])
+                setQuadrilColor(allColors[quadril+1])
+                setSizeBottom(defaultSize)
+                setEditQuadril(quadris[blockValue+quadril])
+            }
+       
+            if (quadril == 4) {
+               return setSizePlusOne(sizePlusOne,bustos[blockValue+busto],cinturas[blockValue+cintura],quadris[blockValue+quadril])
+            }
+            
+               setCinturaDescription(allDescriptions[cintura+1])
+               setCinturaColor(allColors[cintura+1])
+               setEditCintura(cinturas[cintura])
+               setBustoDescription(allDescriptions[busto+1])
+               setBustoColor(allColors[busto+1])
+               setEditBusto(bustos[busto])
+       
+        }
+        //XG
+        if(i===imcRanges.length-2 && imc>imcRanges[i] && imc<imcRanges[i+1]){
+           const blockValue = i
+           const defaultSize = allSizesNames[blockValue]
+           const sizeMinusOne = allSizesNames[blockValue - 1]
+           const bustos = [allSizes[blockValue-1].busto.max,allSizes[blockValue].busto.min,allSizes[blockValue].busto.med-1,allSizes[blockValue].busto.med+1, allSizes[blockValue].busto.max]
+           const cinturas = [allSizes[blockValue-1].cintura.max,allSizes[blockValue].cintura.min,allSizes[blockValue].cintura.med-1,allSizes[blockValue].cintura.med+1, allSizes[blockValue].cintura.max]
+           const quadris = [allSizes[blockValue-1].quadril.max,allSizes[blockValue].quadril.min,allSizes[blockValue].quadril.med-1,allSizes[blockValue].quadril.med+1, allSizes[blockValue].quadril.max]
+           if (quadril == 0) {
+               return setSizeMinusOne(sizeMinusOne, bustos[busto], cinturas[cintura], quadris[quadril])
+           }
+           if (quadril >= 1 && quadril<=4) {
+               //levemente folgado
+               setQuadrilDescription(allDescriptions[quadril+1])
+               setQuadrilColor(allColors[quadril+1])
+               setSizeBottom(defaultSize)
+               setEditQuadril(quadris[quadril])
+           }
+           
+               setCinturaDescription(allDescriptions[cintura+1])
+               setCinturaColor(allColors[cintura+1])
+               setEditCintura(cinturas[cintura])
+               setBustoDescription(allDescriptions[
+                busto+1])
+               setBustoColor(allColors[
+                busto+1])
+               setEditBusto(quadris[
+                busto])
+           
+       }
+       //P ao GG
+       if(i!==imcRanges.length-1 && i!==imcRanges.length-2  && i!==0){
+           if(imc>arr[i] && imc<arr[i+1]){
+               const blockValue = i
+               const defaultSize = allSizesNames[blockValue]
+               const sizePlusOne = allSizesNames[blockValue + 1]
+               const sizeMinusOne = allSizesNames[blockValue - 1]
+               const bustos = [allSizes[blockValue-1].busto.max,allSizes[blockValue].busto.min,allSizes[blockValue].busto.med,allSizes[blockValue].busto.max, allSizes[blockValue+1].busto.min]
+               const cinturas = [allSizes[blockValue-1].cintura.max,allSizes[blockValue].cintura.min,allSizes[blockValue].cintura.med,allSizes[blockValue].cintura.max, allSizes[blockValue+1].cintura.min]
+               const quadris = [allSizes[blockValue-1].quadril.max,allSizes[blockValue].quadril.min,allSizes[blockValue].quadril.med,allSizes[blockValue].quadril.max, allSizes[blockValue+1].quadril.min]
+               if (quadril == 0) {
+                    return setSizeMinusOne(sizeMinusOne, bustos[busto], cinturas[cintura], quadris[quadril])                 
+               }
+               if (quadril >= 1 && quadril<=3 ) {
+                   //levemente folgado
+                   setQuadrilDescription(allDescriptions[busto+1])
+                   setQuadrilColor(allColors[busto+1])
+                   setSizeBottom(defaultSize)
+                   setEditQuadril(bustos[busto])
+               }           
+               if (quadril == 4) {
+                   
+                  return setSizePlusOne(sizePlusOne, bustos[busto], cinturas[cintura], quadris[quadril])
+               }
+             
+                 setCinturaDescription(allDescriptions[cintura+1])
+                 setCinturaColor(allColors[cintura+1])
+                 setEditCintura(cinturas[cintura])
+                 setBustoDescription(allDescriptions[quadril+1])
+                 setBustoColor(allColors[quadril+1])
+                 setEditBusto(bustos[busto])
+                   }
+           }
+       })
 
+
+
+
+    }
+    const chooseBestVestidoDoll = (imc:number,busto:number,cintura:number,quadril:number) =>{   
+        imcCheck(imc)
+        const setSizePlusOne = (sizePlusOne:string,apxBusto:number,apxCintura:number,apxQuadril:number) =>{
+            setBustoDescription(allDescriptions[busto-2])
+            setBustoColor(allColors[busto-2])
+            setEditBusto(apxBusto)
+            setCinturaDescription(allDescriptions[cintura])
+            setCinturaColor(allColors[cintura])
+            setEditCintura(apxCintura)
+            setQuadrilDescription(allDescriptions[quadril])
+            setQuadrilColor(allColors[quadril])
+            setEditQuadril(apxQuadril)
+            setSizeBottom(sizePlusOne)
+        }
+        const setSizeMinusOne = (sizeMinusOne:string,apxBusto:number,apxCintura:number,apxQuadril:number) =>{
+            setBustoDescription(allDescriptions[busto+5])
+            setBustoColor(allColors[busto+5])
+            setEditBusto(apxBusto)
+            setCinturaDescription(allDescriptions[cintura])
+            setCinturaColor(allColors[cintura])
+            setEditCintura(apxCintura)
+            setQuadrilDescription(allDescriptions[quadril])
+            setQuadrilColor(allColors[quadril])
+            setEditQuadril(apxQuadril)
+            setSizeBottom(sizeMinusOne)
+        }
+
+        imcRanges.forEach((v,i,arr)=>{
+            if(i===imcRanges.length && i ===imcRanges.length-1)return
+            //PP
+           if(i===0 && imc>imcRanges[i] && imc<imcRanges[i+1]){
+           console.log('your size is PP')
+            const bustos = [allSizes[i].busto.min, allSizes[i].busto.min+1, allSizes[i].busto.med+1, allSizes[i].busto.max, allSizes[i+1].busto.min]
+            const quadris = [allSizes[i].quadril.min, allSizes[i].quadril.min+1, allSizes[i].quadril.med+1, allSizes[i].quadril.max, allSizes[i+1].quadril.min]
+            const cinturas = [allSizes[i].cintura.min, allSizes[i].cintura.min+1, allSizes[i].cintura.med+1, allSizes[i].cintura.max, allSizes[i+1].cintura.min]
+            
+            const blockValue = i
+            const defaultSize = allSizesNames[blockValue]
+            const sizePlusOne = allSizesNames[blockValue + 1]
+            console.warn('is 0?')
+            console.warn(busto, typeof busto,' busto log warn')
+            
+            if (quadril>=0 && quadril<4) {
+                setQuadrilDescription(allDescriptions[quadril+1])
+                setQuadrilColor(allColors[quadril+1])
+                setSizeWhole(defaultSize)
+                setEditQuadril(quadris[blockValue+quadril])
+            }
+       
+            if (quadril == 4) {
+               return setSizePlusOne(sizePlusOne,bustos[blockValue+busto],cinturas[blockValue+cintura],quadris[blockValue+quadril])
+            }
+            
+               setCinturaDescription(allDescriptions[cintura+1])
+               setCinturaColor(allColors[cintura+1])
+               setEditCintura(cinturas[cintura])
+               setBustoDescription(allDescriptions[busto+1])
+               setBustoColor(allColors[busto+1])
+               setEditBusto(bustos[busto])
+       
+        }
+        //XG
+        if(i===imcRanges.length-2 && imc>imcRanges[i] && imc<imcRanges[i+1]){
+           const blockValue = i
+           const defaultSize = allSizesNames[blockValue]
+           const sizeMinusOne = allSizesNames[blockValue - 1]
+           const bustos = [allSizes[blockValue-1].busto.max,allSizes[blockValue].busto.min,allSizes[blockValue].busto.med-1,allSizes[blockValue].busto.med+1, allSizes[blockValue].busto.max]
+           const cinturas = [allSizes[blockValue-1].cintura.max,allSizes[blockValue].cintura.min,allSizes[blockValue].cintura.med-1,allSizes[blockValue].cintura.med+1, allSizes[blockValue].cintura.max]
+           const quadris = [allSizes[blockValue-1].quadril.max,allSizes[blockValue].quadril.min,allSizes[blockValue].quadril.med-1,allSizes[blockValue].quadril.med+1, allSizes[blockValue].quadril.max]
+           if (quadril == 0) {
+               return setSizeMinusOne(sizeMinusOne, bustos[busto], cinturas[cintura], quadris[quadril])
+           }
+           if (quadril >= 1 && quadril<=4) {
+               //levemente folgado
+               setQuadrilDescription(allDescriptions[quadril+1])
+               setQuadrilColor(allColors[quadril+1])
+               setSizeWhole(defaultSize)
+               setEditQuadril(quadris[quadril])
+           }
+           
+               setCinturaDescription(allDescriptions[cintura+1])
+               setCinturaColor(allColors[cintura+1])
+               setEditCintura(cinturas[cintura])
+               setBustoDescription(allDescriptions[
+                busto+1])
+               setBustoColor(allColors[
+                busto+1])
+               setEditBusto(quadris[
+                busto])
+           
+       }
+       //P ao GG
+       if(i!==imcRanges.length-1 && i!==imcRanges.length-2  && i!==0){
+           if(imc>arr[i] && imc<arr[i+1]){
+               const blockValue = i
+               const defaultSize = allSizesNames[blockValue]
+               const sizePlusOne = allSizesNames[blockValue + 1]
+               const sizeMinusOne = allSizesNames[blockValue - 1]
+               const bustos = [allSizes[blockValue-1].busto.max,allSizes[blockValue].busto.min,allSizes[blockValue].busto.med,allSizes[blockValue].busto.max, allSizes[blockValue+1].busto.min]
+               const cinturas = [allSizes[blockValue-1].cintura.max,allSizes[blockValue].cintura.min,allSizes[blockValue].cintura.med,allSizes[blockValue].cintura.max, allSizes[blockValue+1].cintura.min]
+               const quadris = [allSizes[blockValue-1].quadril.max,allSizes[blockValue].quadril.min,allSizes[blockValue].quadril.med,allSizes[blockValue].quadril.max, allSizes[blockValue+1].quadril.min]
+               if (quadril == 0) {
+                    return setSizeMinusOne(sizeMinusOne, bustos[busto], cinturas[cintura], quadris[quadril])                 
+               }
+               if (quadril >= 1 && quadril<=3 ) {
+                   //levemente folgado
+                   setQuadrilDescription(allDescriptions[busto+1])
+                   setQuadrilColor(allColors[busto+1])
+                   setSizeWhole(defaultSize)
+                   setEditQuadril(bustos[busto])
+               }           
+               if (quadril == 4) {
+                   
+                  return setSizePlusOne(sizePlusOne, bustos[busto], cinturas[cintura], quadris[quadril])
+               }
+             
+                 setCinturaDescription(allDescriptions[cintura+1])
+                 setCinturaColor(allColors[cintura+1])
+                 setEditCintura(cinturas[cintura])
+                 setBustoDescription(allDescriptions[quadril+1])
+                 setBustoColor(allColors[quadril+1])
+                 setEditBusto(bustos[busto])
+                   }
+           }
+       })
+    }
+    //doll
+
+
+    const fromEditSetter = (sizeTop:string,sizeBottom:string,sizeWhole:string,bustoCm:number,cinturaCm:number,quadrilCm:number,bustoDescription:string,bustoColor:string,cinturaDescription:string,cinturaColor:string,quadrilDescription:string,quadrilColor:string) =>{
+
+        setSizeTop(sizeTop);
+        setSizeBottom(sizeBottom);
+        setSizeWhole(sizeWhole);
+
+        // console.log('router query inside the setter', router.query)
+
+        setEditBusto(bustoCm)
+        setEditCintura(cinturaCm)
+        setEditQuadril(quadrilCm)
+
+        setBustoDescription(bustoDescription)
+        setBustoColor(bustoColor)
+        setCinturaDescription(cinturaDescription)
+        setCinturaColor(cinturaColor)
+        setQuadrilDescription(quadrilDescription)
+        setQuadrilColor(quadrilColor)
+    }
+
+
+    const resultMasterReturn = (categoria: string, doll:any, edit:any) =>{
+    if(categoria && doll || edit){
+        // console.log('router loads', doll,'dollvalue', )
+        if(doll){
+            // console.log('doll is true')
+            allPossibleCategorias.map((v,i,arr)=>{
+                if(categoria===v){
+                    console.log(categoria,'categoria is equal' ,v )
+                    functionsDollObj[i](imc,bustoDoll,cinturaDoll,quadrilDoll)
+                  
+                }
+            })
+            
+        }
+        if(edit){
+            fromEditSetter(String(router.query.sizeTop),String(router.query.sizeBottom),String(router.query.sizeWhole),Number(router.query.bustoCm),Number(router.query.cinturaCm),Number(router.query.quadrilCm),String(router.query.bustoDescription),String(router.query.bustoColor),String(router.query.cinturaDescription),String(router.query.cinturaColor),String(router.query.quadrilDescription),String(router.query.quadrilColor));
+        }
+    }
+    else return null
+    
+
+
+    
+    }
+
+useEffect(()=>{
+    resultMasterReturn(categoria,doll,edit)
+},[])
+
+const factory = () =>{
+    if(sizeTop|| sizeBottom || sizeWhole){
+
+        const factoryResult:any[] = []
+        allSizesNames.forEach((v,i,arr)=>[
+            factoryResult[i] = {}
+        ])
+        
+    let originalSize  = ''
+    allPossibleCategorias.forEach((v,i,arr)=>{
+        if(categoria===v){
+            originalSize = sizes[i]
+            console.log(allSizesNames.indexOf(sizes[i]), 'indexofsizes[i')
+            // const a =  Object.values(factoryInfo)
+            factoryResult[allSizesNames.indexOf(sizes[i])] = factoryInfo 
+
+        }
+    })
+    console.log(factoryResult,'this is factory restul')
+    //ate aqui conseguimos passar pro index correto a opção inicialmente fabricada
+    
+    factoryResult.forEach((v,i,arr)=>{
+        if(i !== allSizesNames.indexOf(originalSize)){
+            const baseSize = Object.assign({},factoryResult[allSizesNames.indexOf(originalSize)])
+            
+            // console.log('this is baseSize', baseSize)
+            
+            const sizeDifference = i-allSizesNames.indexOf(originalSize)
+            // console.log('this is the size difference?', sizeDifference)
+            
+            
+            baseSize.bustoDescription = baseSize.bustoDescription-sizeDifference*2 
+            baseSize.cinturaDescription = baseSize.cinturaDescription-sizeDifference*2 
+            baseSize.quadrilDescription = baseSize.quadrilDescription-sizeDifference*2 
+            allPossibleCategorias.forEach((v,index,arr)=>{
+                if(categoria===v){
+                    baseSize[resultNameCategorias[index]] = allSizesNames[i]
+                }
+            })
+            
+            for (const key in baseSize) {
+                if (baseSize.hasOwnProperty(key)) {
+                    //   console.log(`${key}: ${baseSize[key]}`);
+                    if(baseSize[key] && baseSize[key]>=7 && baseSize[key]<16){
+                        baseSize[key]=6
+                    }
+                    if(baseSize[key] && baseSize[key]<0){
+                        baseSize[key]=0
+                    }
+                    // console.log(`${key}: ${baseSize[key]}`);
+                }
+              }
+            //   console.log('after the bugfix with the indexes of description and color, now baseSize is like this')
+              const indexOfFutureSize = i
+              baseSize.editBusto = allSizes[indexOfFutureSize].busto.med
+              baseSize.editCintura = allSizes[indexOfFutureSize].cintura.med
+              baseSize.editQuadril = allSizes[indexOfFutureSize].quadril.med
+            //   console.log(baseSize, indexOfFutureSize,'indexofnextsize')
+            const result = [baseSize.sizeTop,baseSize.sizeBottom,baseSize.sizeWhole,baseSize.editBusto,baseSize.editCintura,baseSize.editQuadril,allDescriptions[baseSize.bustoDescription], allColors[baseSize.bustoDescription], allDescriptions[baseSize.cinturaDescription], allColors[baseSize.cinturaDescription], allDescriptions[baseSize.quadrilDescription],allColors[baseSize.quadrilDescription]]
+            console.log(result, 'log of result inside loop')
+            factoryResult[i] = result
+        }
+    })
+    //wiping the original object to an
+    console.log(factoryResult,'this is factory restul')
+    factoryResult[allSizesNames.indexOf(originalSize)] = [factoryInfo.sizeTop,factoryInfo.sizeBottom,factoryInfo.sizeWhole, factoryInfo.editBusto,factoryInfo.editCintura,factoryInfo.editQuadril, allDescriptions[factoryInfo.bustoDescription], allColors[factoryInfo.bustoDescription], allDescriptions[factoryInfo.cinturaDescription], allColors[factoryInfo.cinturaDescription], allDescriptions[factoryInfo.quadrilDescription],allColors[factoryInfo.quadrilDescription]]
+    console.log(factoryResult,'this is factory restul after orignial value changes')
+    setFactoryResultState(Object.assign({},factoryResult))
+    
+
+}
+}
+
+useEffect(()=>{
+if(!factoryResultState)
+    factory()
+})
+
+const changeSize = (index:number) =>{
+    if(factoryResultState && Array.isArray(factoryResultState[index])){
+        console.log('going to be this>', factoryResultState[index], 'at index', index)
+        fromEditSetter(...factoryResultState[index] as [string,string,string,number,number,number,string,string,string,string,string,string])
+    }
+}
+
+
+if(!router.isFallback){
+    return (<div className={` ${roboto.className} bg-white flex justify-start outline-1 rounded-lg shadow-md items-center`}>
+
+        <div className='max-w-[19rem] min-w-[19rem] self-start'>
+        {/* configurar isso pra ser gerado tbm, exemplo, calça nao precisa de busto */}
+         {sizeTop || sizeWhole? <SugarBustoSVG bustoColor={bustoColor} /> :null} 
+         <SugarCinturaSVG cinturaColor={cinturaColor} />
+         <SugarQuadrilSVG quadrilColor={quadrilColor} />
+         {/* configurar isso pra ser gerado tbm */}
+
+        <img alt={'doll std img'} className="object-cover object-top h-[26rem] w-96" src='/imgs_lela/030303.jpeg'/>
+        <ChangeSizeButtons baseSize={{sizeTop,sizeBottom,sizeWhole}} changeSize={(i:number)=>changeSize(i)}></ChangeSizeButtons>
+        {/* <OtherSizesFactoryR baseSize={factoryInfo} setter={(a: string,b: string,c: string,d: number,e: number,f: number,g: string,h: string,i: string,j: string,k: string,l: string)=>{fromEditSetter(a,b,c,d,e,f,g,h,i,j,k,l)}}></OtherSizesFactoryR> */}
+
+    </div>
+    <div className=''>
+        <h1 className='text-gray-800 font font-medium text-xl py-4'> Sua opção de {categoria} é {sizeTop || sizeBottom || sizeWhole}</h1>
+        
+        {sizeTop || sizeWhole?<div className='mt-16 text-black'>Busto:{bustoDescription}</div> :<div className="mt-10">-</div>} 
+        <div className='mt-8 text-black'>Cintura:{cinturaDescription}</div>
+        <div className='mt-16 text-black'>Quadril:{quadrilDescription}</div>
+
+    <div className='flex flex-col justify-start mt-8'>
+         <button className="rounded-lg bg-white text-black shadow-lg py-2 my-2 " onClick={() => router.push({ pathname: `${window.location.origin}/${encodeURIComponent(encodedImgUrl)}`, query:{categoria:categoria}})}>Reiniciar</button>
+        {doll ? <button className="rounded-lg bg-white text-black shadow-lg py-2 my-2 " onClick={() => router.push({ pathname: `${window.location.origin}/doll/DollPage`, query: { imc: imc, encodedImgUrl: encodedImgUrl, categoria:categoria } })}>Voltar</button> : null}
+        <button className="rounded-lg bg-white text-black shadow-lg py-2 my-2 " onClick={() => {router.push({ pathname: `${window.location.origin}/medidas/EditarMedidas`, query: { imc: imc, editBusto: editBusto, editCintura: editCintura, editQuadril: editQuadril, encodedImgUrl: encodedImgUrl, categoria:categoria } }) ; console.log({editBusto,editCintura,editQuadril})}}>Editar medidas</button>
+         {/* <button className="rounded-lg bg-white text-black shadow-lg py-2 my-2 " onClick={()=>router.push({pathname:'/medidas/NotFound', query:{imc:imc, encodedImgUrl:encodedImgUrl}})}>NotFound</button> */}
+         <a className='rounded-lg bg-white text-black shadow-lg py-2 my-2 text-center' href='/files/fita-metrica-eufloria.pdf' target={"_blank"} rel='noopener noreferrer'>
+           Fita Metrica
+            </a>
         </div>
+
+    </div>
 </div>)
+}
+else return <div>loading...</div>
 }
